+++ api/server/structs/BUILD.bazel
@@ -7,6 +7,7 @@ go_library(
         "conversions.go",
         "conversions_blob.go",
         "conversions_block.go",
+        "conversions_execution.go",
         "conversions_lightclient.go",
         "conversions_state.go",
         "endpoints_beacon.go",
@@ -19,6 +20,7 @@ go_library(
         "endpoints_node.go",
         "endpoints_rewards.go",
         "endpoints_validator.go",
+        "execution.go",
         "other.go",
         "state.go",
     ],
